refactor: move mapper function from mapper file to dto file and remove mapper file
