+++ trends/src/main/kotlin/net/thechance/trends/api/dto/category/CategoryResponse.kt
@@ -1,5 +1,6 @@
 package net.thechance.trends.api.dto.category
 
+import net.thechance.trends.entity.Category
 import java.util.UUID
 
 data class CategoryResponse(
@@ -7,3 +8,11 @@ data class CategoryResponse(
     val name: String,
     val emoji: String
 )
+
+fun Category.toCategoryResponse(): CategoryResponse {
+    return CategoryResponse(
+        id = this.id,
+        name = this.name,
+        emoji = this.emoji
+    )
+}
